+++ app/Ninja/Repositories/InvoiceRepository.php
@@ -494,13 +494,15 @@ class InvoiceRepository extends BaseRepository
             }
         }
 
-        foreach ($invoice->documents as $document){
-            if(!in_array($document->public_id, $document_ids)){
-                // Removed
-                // Not checking permissions; deleting a document is just editing the invoice
-                if($document->invoice_id == $invoice->id){
-                    // Make sure the document isn't on a clone
-                    $document->delete();
+        if ( ! $invoice->wasRecentlyCreated) {
+            foreach ($invoice->documents as $document){
+                if(!in_array($document->public_id, $document_ids)){
+                    // Removed
+                    // Not checking permissions; deleting a document is just editing the invoice
+                    if($document->invoice_id == $invoice->id){
+                        // Make sure the document isn't on a clone
+                        $document->delete();
+                    }
                 }
             }
         }
